requirements/main.txt
numpy>=1.21.6
Pillow>=9.2.0
tqdm>=4.64.0
matplotlib>=3.2.2
scipy>=1.4.1
albumentations>=1.0.3
pandas>=1.1.4
torchsampler
comet-ml
timm
wheel
setuptools
ninja
nvidia-pyindex
